+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Serialization of local components in Blackboard Container.
+- Blackboard.GetValueType() method that returns a type of a property by its property name.
+- Blackboard.GetPropertyNames() method that returns all property names of properties contained in Blackboard.
 
 ## [1.1.0] - 2020-09-28